+++ NEWS.md
@@ -3,6 +3,9 @@
 * `unnest()` will run with named and unnamed list-columns of same length
   (@hlendway, #460).
 
+* `crossing()` (and hence `expand()`) now works with list-columns 
+  (#446, @SamanthaToet).
+
 * `crossing()` preserves factor levels (#410).
 
 # tidyr 0.8.2