+++ NEWS.md
@@ -1,5 +1,7 @@
 # tidyr 0.8.2.9000
 
+* `unnest()` preserves column being unnested when input is zero-length (#483).
+
 * `unnest()` will run with named and unnamed list-columns of same length
   (@hlendway, #460).