+++ NEWS.md
@@ -1,4 +1,4 @@
-# tidyr 0.8.2.9000
+# tidyr 0.8.3
 
 * `crossing()` preserves factor levels (#410), now works with list-columns 
   (#446, @SamanthaToet). (These also help `expand()` which is built on top